backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
psycopg2-binary==2.9.11
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pillow==11.0.0
opencv-python==4.10.0.84
numpy==2.2.0
websockets==14.1
aiofiles==25.1.0
httpx==0.28.1
redis==7.1.0
celery==5.4.0
python-dotenv==1.0.1
loguru==0.7.3
prometheus-client==0.21.1
motor==3.6.0
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
isort==5.13.2
aiosqlite==0.21.0
email-validator==2.3.0
openai==1.57.2
groq==0.13.0